+++ public/admin/_app/immutable/nodes/14.U2EnF1IG.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/BahMn3972.js";export{e as component};