+++ public/admin/_app/immutable/nodes/19.CuQutSTr.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/BJ4Izu1K2.js";export{e as component};