+++ public/admin/_app/immutable/nodes/22.Bd34zhFC.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/N9jkHQk8.js";export{e as component};