+++ public/admin/_app/immutable/nodes/23.DdrA0JNR.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/CUB3-Qy6.js";export{e as component};