+++ public/admin/_app/immutable/nodes/24.DiwyjuXa.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/D5cWeMI_.js";export{e as component};